修改16位匹配过程中断的bug,
修改16位单波段的整个流程运行BUG，
并修改了大尺度下添加同名点点纠正会退出的BUG

+++ DOMRectify/MainFrm.h
@@ -22,6 +22,7 @@
 #include "DOMRectifyPrj.h"
 #include "GradientProgressCtrl.h"
 #include "DomRectifyPro.h"
+
 class CChildFrame;
 class CDlgPrjNew;
 class CDlgPrjNew2;
@@ -47,7 +48,8 @@ public:
 	bool m_bAutoSelectRef;
 	bool m_bViewPoint;
 	bool m_bAutoRectify;
-	void ReadImage(CString strImagePath, int stcol, int strow, int edcol, int edrow, int memWidth, int memHeight, BYTE*&data);
+	//void ReadImage(CString strImagePath, int stcol, int strow, int edcol, int edrow, int memWidth, int memHeight, BYTE*&data);
+	void ReadImage(CString strImagePath, int stcol, int strow, int edcol, int edrow, int memWidth, int memHeight, unsigned int*&data);
 	void SaveImage(CString strImagePath, int nCols, int nRows, int nBands, BYTE*data, const char*pszFormat);
 
 	// Ьиад
